Prevent complete DOM update for breakpoints sidebar when debugging
This CL introduces partial list update mechanism when the number of
breakpoints in the sidebar panel remains the same. In this case,
the component will detect if there were any changes in the breakpoints
and it will update only the changed breakpoints. This change speeds
up the step navigation in the debugger.

Before: https://i.imgur.com/8ed4258.png
After: https://i.imgur.com/aveakGL.png

Bug: 1069694
Change-Id: Iff420b4edef7b3505d2e3662c6bea261f970dd34