Restore old behavior in ResourceTreeModel
The CL https://crrev.com/c/1934208 caused failed CDP messages to not
resolve its Promise with null, but instead propagate the exception
via "reject". This caused the ResourceTreeModel not to properly
intialize and, cascading down, the ConsoleModel to swallow messages.

This CL fixes this by adding a catch handler that calls
{_processCachedResources} with null, thus, restoring the old behavior.

R=changhaohan@chromium.org

Bug: chromium:1063012
Change-Id: I16da2cf2a77e22493e671a01ef1dfed222b2f1d9