[issues] Report issue number instead of aggregated issue number
This change reports the number of issues known to DevTools instead of
the number of aggregated issues. This means that any improvement will
result in a decreased number. We think this provides better feedback for
developers.

To make this work properly, this change also introduces de-duplication
of issues, as cookie issues are ofen reported more than once by the
chromium back-end.

Bug: chromium:1057979
Change-Id: Iebea90f3840a5c782866713990b03ae47712346b