Update patch set 5
Patch Set 5:

Ok, thought about tests. I'll take it back, we could actually implement an e2e test.

To test this, you could write an e2e test that tries to set a breakpoint on an unbreakable line and a breakable line. For checking whether breaking works on the breakable line, see this test:

https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/test/e2e/sources/debug-raw-wasm_test.ts;l=12?q=debug-raw-wasm_test.ts&ss=chromium&originalUrl=https:%2F%2Fcs.chromium.org%2F

For checking that the line is not breakable, you could use the same code and check for the return value of addBreakpointForLine.

For the visuals, I'm currently thinking if we can do this: get all instances of '.Codemirror-line', similar to here with '.Codemirror-linenumber'; get the ith one (whichever is not breakable), and get its parent node (there is a helper function for that), and check that it's the class is 'cm-line-without-source-mapping' (or whatever we are going to use).

https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/test/e2e/helpers/sources-helpers.ts?q=addBreakpointForLine&ss=chromium&originalUrl=https:%2F%2Fcs.chromium.org%2F

However, I'm not sure the last part is suitable for an e2e test. A visual comparison would be better here, but I'm not sure our infrastructure is there yet.

Patch-set: 5
Reviewer: Gerrit User 1396836 <1396836@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0