Move `UIString` into `Platform`
Platform is the "lowest" part of DevTools - e.g. it depends on nothing.

We have code we want to move to Platform but cannot because it needs
`ls` and therefore creates a circular dependency. So this CL moves
UIString into Platform.

To avoid rewriting a lot of imports we re-export it from `common.js`.

Change-Id: I399bbe8593e043148c4c7e51f598cbd73a89f639