[issues] Add affected sources to Issue view
This change adds 'sources' as an affected resource for issues.
If provided the sources will render as a linkified file name with line
number:

      https://cdn.foo.example/js/my-script.js -> script.js:5

Activating the link will open the sources panel to the line and column
number of the affected source (note that only the line number is displayed).

The affected sources will be useful for future issues such as [deprecation
of synchronous XMLHttpRequest on the main thread][1].

[1]: https://imgur.com/MYpi6cY

The below images are using dummy data found in patchset 2 and are for
example purposes only.

Screenshot: https://imgur.com/6Xiqs75
Gif: https://imgur.com/NY92K1w
Keyboard accessibility gif: https://imgur.com/fsG2ZhT
Bug: chromium:1070376
Change-Id: I37906f24e731cffff5da59a6e0084df1a2413732