Fix placeholder height for live console expressions
It turns out we are poking into CodeMirror internals and implicitly
relying on its styling. In https://github.com/codemirror/CodeMirror/pull/5963
a selector on a `pre` element in CodeMirror was updated to be more
specific. We were implicitly relying on this styling, as we inject
`pre` elements as placeholders for these console live expressions.

Therefore, add these classes explicitly to make sure the height
is correct. For future changes, we should remove the poking of
internals and also more frequently update CodeMirror so that
it is easier to bisect what the culprit is.

R=bmeurer@chromium.org

Fixed: 1074197
Change-Id: Ia49e4d97605b5fcd6c17b83d124a3f39a42a294c