Update patch set 13
Patch Set 13:

Hi John, I'd be happy to include resources of this type in the issues panel. However, it would be nice if we had an issue in the back-end that reports this new kind of a affected resource. Is there a CL in flight?

Patch-set: 13
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>