Improve readability of colorpicker output
This patch makes the colorpicker output in various formats more
readable by using appropriate unit suffixes where applicable.

For any alpha values, we now use `%` notation since that makes the
range clear (0-100%).

    rgb(0 0 0 / 0.5)
    →→→→→→→→→→→→→→→→
    rgb(0 0 0 / 50%)

For any values that can be expressed in degrees, we were previously
outputting the degrees as a number. This patch adds the `deg` suffix
to make it more clear that the range is 0-360deg.

    hsl(288 92% 31% / 87%)
    →→→→→→→→→→→→→→→→→→→→→→→→→
    hsl(288deg 92% 31% / 87%)

This simplifies the code as well: we can now handle the HSL/HSLA and
RGB/RGBA formats in a single `case`, with just a conditional for the
alpha, as opposed to handling all these cases separately.

Bug: chromium:1075869
Change-Id: I2afad7ccce95d14cbe15e7bf4668281335152bdb