Update patch set 3
Patch Set 3:

> Patch Set 1: Code-Review+1
> 
> (6 comments)
> 
> Overall lgtm % some comments.
> 
> I am a bit torn because of the high level of decoupling between the IssuesManager and the IssuesAggregator and that they only communicate via events with each other.
> 
> On the one hand, it might make the code easier to read if IssuesManager is the only "event source" and IssuesManager <=> IssuesAggregator just call methods directly on each other. I can imagine any more users that would want their own IssuesAggregator. On the other hand, this approach makes the code more testable, which is very nice.

Yes, I share that view. One design consideration was that I did not want the IssuesManager to know about the IssuesAggregator. An option could be an observer pattern. That would also have the advantage that the event dispatch becomes a function call that can actually be type-checked.

Patch-set: 3