Update patch set 14
Patch Set 14:

> Patch Set 14: Commit-Queue+1
> 
> (3 comments)
> 
> > Patch Set 13:
> > 
> > Hi John, I'd be happy to include resources of this type in the issues panel. However, it would be nice if we had an issue in the back-end that reports this new kind of a affected resource. Is there a CL in flight?
> 
> Thanks for considering it!
> 
> I do not have an implementation for a back-end issue yet, I am hoping you will take this front-end portion first.
> 
> I definitely want to make this useful for futures issues we add such as the deprecation warning I suggested in the CL description and affected sources are demoed in the webhint explainer [1][2].
> 
> [1]: https://imgur.com/nbXgjJp
> [2]: https://docs.google.com/document/d/1IHoVXF5tn7kD8UaSIRPuIvEn5cHdjlyIPEvqEQiIv0k/edit?usp=sharing

Hi John, I agree that this affected resource will be useful, and not only for webhints but also for other issues, potentially.

Have you considered adding an end-to-end test that exercises this code (since it doesn't seem to be exercised anywhere else at the moment)? 

I'm not opposed to landing this as-is, and LGTM-wise you are good-to-go.

Patch-set: 14
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>