Check in CodeMirror 5.53.2 in front_end/third_party
We already include CodeMirror in the front_end/ (in cm/, cm_headless/,
cm_modes/ and cm_web_modes/), which does not follow the third_party
guidelines of Chromium. Moreover, it includes the build output
from CodeMirror, which can't be used with TypeScript. Instead,
we include the sources (which use ES modules) and will use
`@types/codemirror` in a follow-up CL to make it work with TypeScript.

The BUILD.gn rules currently copy the code to the correct `gen/` folder
as well as `resources/inspector`. However, it does not yet include
the files in the `grd` file, to prevent code duplication. That will happen
in a follow-up change.

R=mathias@chromium.org,bmeurer@chromium.org

Bug: 1011811, 1076825
Change-Id: I25f9b230f88262ed008f96f445ecf2e4240790d1