Update patch set 2
Patch Set 2:

(1 comment)

I tried it, but it is not scrolling the view to fit tree items that are outside the current view.

Go to sources, right click on a tab and select 'Reveal in sidebar', if item is not present in view, view is not scrolled.
Repro: https://imgur.com/a/YmU8YbL

Patch-set: 2
Reviewer: Gerrit User 1341877 <1341877@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>