Remove incorrect source from CodeMirror BUILD.gn copy action
This source is a directory, not a file. This caused crashes on the
Chromium bots and closed our trees. It is unclear why this crash
did not happen locally nor on our CQ. I will follow up with the
GN implementation of `copy` to prevent copying directories.

R=aerotwist@chromium.org

No-Tree-Checks: true
Fixed: 1077186
Bug: 1076825, 1011811
Change-Id: I564a0d01ea98f2e07c8f9f94d2d7e3628e484a92