[issues] Add affected elements
This changes added the element(s) affected resource type. The an affected
element represented by a subset of Protocol.DOM.BackendNode properties [1].

In the issues panel view the `backendNodeId` and `nodeName` are used to
display a deferred node link that will only push the node to the front-end
when activated by the user.

Activating the tag name link opens the elements panel and highlights the
node.

Notes that only the element tag is displayed and not attributes on the element
such as id or classes.

This change is only the front-end displaying of the affected element and
will be useful for future back-end issue implementations, e.g. _Preload
warnings_ would link to the offending <script> tag in the Elements panel [2].

[1]: https://chromedevtools.github.io/devtools-protocol/tot/DOM/#type-BackendNode
[2]: https://imgur.com/nAMHsfm

Screenshot: TDB
Keyboard Accessibility GIF: TBD
Bug: chromium:1070375
Change-Id: Ia27a067c92bdf5fb14bb008151c4d80293d70678