[issues] Improve linking cookies -> issues
The cookies pane has a mysterious filter "Only blocked" that
filters cookies based on whether there is an associated issue.
This CL renames the filter to mention that fact, and adds a
context menu option to reveal one of the associated issues.

As a bonus, it includes some TypeScriptification fixes.

Bug: chromium:1057979
Change-Id: Idb557c4c6488893948466fa9ba70ecb94457bbe9