Restore ShortcutRegistry Event Consumption
This CL fixes a regression where ShortcutRegistry allows keypresses
to bubble up to the browser when a dialog is open, which has the effect
of opening the Chromium help page if you hit F1 while the DevTools
settings are open. The cause was a change to when events are consumed in
this chords CL [1] that combined the check for Dialog.hasInstance() with
other failure conditions which didn't require consuming the event.

[1]
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2125620

Bug: 1077667
Change-Id: I10d15a9d2aa967e3849804bc901de49a8a8683be