Move issue indicator triangle to the left
This CL moves the triangle indicator for expanding/collapsing issues
to the left side of the line. We received feedback that the UI does not
clearly show that the issue headline is clickable. We hope that the
triangle indicator now makes this more obvious.

Screenshot: https://imgur.com/iw48RNT

Bug: chromium:1057979
Change-Id: I46474b0bcc5bd1e9547c54607c9fcdff654eb779