[issues] Add affected elements
This changes added the element(s) affected resource type. The an affected
element represented by a subset of Protocol.DOM.BackendNode properties [1].

In the issues panel view the `backendNodeId` and `nodeName` are used to
display a deferred node link that will only push the node to the front-end
when activated by the user.

Activating the tag name link opens the elements panel and highlights the
node.

Note that only the element tag is displayed and not attributes on the element
such as id or classes.

This change is only the front-end displaying of the affected element and
will be useful for future back-end issue implementations, e.g. _Preload
warnings_ would link to the offending <script> tag in the Elements panel [2].

The below screenshot and gif are demo's using dummy data in patchset 3 [3], unfortunately
the backEndNodId changes often so there is no guarantee the dummy data
will work correctly on every page.

[1]: https://chromedevtools.github.io/devtools-protocol/tot/DOM/#type-BackendNode
[2]: https://imgur.com/nAMHsfm
[3]: https://crrev.com/c/2176406/2..3

Screenshot: https://imgur.com/oG78y7O
Keyboard Accessibility GIF: https://imgur.com/dMloRDc
Bug: chromium:1070375
Change-Id: Ia27a067c92bdf5fb14bb008151c4d80293d70678