Update patch set 6
Patch Set 6: Code-Review+1

> Patch Set 6: Code-Review+1
> 
> (1 comment)
> 
> LGTM. Just one minor typo I noticed.
> 
> Sigurd/Simon, do either of you know if there's a reasonable way to plumb a fake issue into an end-to-end test? I poked around a bit but didn't see anything obvious. I know we didn't require this for the "Affected Sources" change so I don't think it's blocking. That said it would be nice to have something that exercises this code until we have concrete issues using it if possible.

No, there is currently no official way, we would need to talk to architecture folks. That being said, we would strongly like to keep end-to-end test truly end-to-end. For other affected resources we have some kind of integration test where we mocked the backend to test that the middle layers work, but not the UI. This is not possible since we don't report affected elements yet from the backend and affected elements are also not wired up with the rest of issues machinery.

Side note: We should be able to have standalone tests for this particular case once we move towards web components, since we could instantiate a standalone "AffectedElement" view and unit test it.

lgtm on this particular CL. If you plan to add more frontend features around the issues panel that will go unused for a bit, we would prefer if you could put together a small Design Doc that shows the bigger picture here (what backend changes are planed and how do they fit with these fronted updates).

 Thanks for your work on this!

Patch-set: 6
Reviewer: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1