Update patch set 1
Patch Set 1:

I wonder if it would be better to change SettingsScreen.js so it only gets the settings from the explicit list of categories it already has. The problem with relying on specific properties being present in module.json is that it's not very obvious what the effect is, and so we might make the same mistake in the future.

Having SettingsScreen.js as the source of truth for what appears in the settings screen seems ideal, IMHO. WDYT?

Patch-set: 1
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>