Add sentence about running single tests in the e2e runner
Previously we had the `--test-file` flag to run a single test.
However, since we now use Mocha as the runner, we can use the
`{describe,it}.only` pattern once again. (This previously didn't
work as we were running tests in a worker, which broke Mocha's
test filtering mechanism. This has since been fixed.)

R=joselea@microsoft.com

Change-Id: Iedb90b034cd642cd617f6a51d219d6e40438846a