Update patch set 1
Patch Set 1:

Hi Simon, this is the CL that removed the hack again (dependent on https://crrev.com/c/2183937 being landed + rolled to us).

I'm looking for a rubber-stamp +1 here, we can discuss the code in the follow up that actually uses the new data, but here my goal is to get rid of the hack.

Patch-set: 1
Reviewer: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false