Focus request details tab when revealed
This CL fixes an accessibility issue in the network panel where the
details pane appears without taking focus or otherwise revealing its
existence to screen reader users. The fix is to focus the details pane
when it's revealed by invoking an item in the datagrid.

Video: https://i.imgur.com/jJDhT2k.mp4

Bug: 963183
Change-Id: Ib90da5bacec59748e19b1d3788c884c8378dd24f