Forbid adding blank entries in DataGrid
Blank key value pairs can currently be added in local and session storage which leads to the ui displaying two empty entries instead of one.
Blank entries can currently be added because comparison of newText (empty string) and the previous value (null) evaluates to true in DataGridImpl._editingCommitted.
This PR changes this by setting the previous value (valueBeforeEditing) to empty string in case it is null.

Fixed: 1077736
Change-Id: I38961e7e744e1abf41b2e36fdaaee76ed0abcbe2