Add css rules for navigation buttons when focused in screencast.
With regard the bug described below, we noticed that when the focus is
supposed to be on the said buttons, the button has focus but there is
no visual indication. We add a css rule so that the focus is shown the
same way as being hovered over.

Images after (from navigation url, use shift+tab, to reload and back
buttons):




Accessibility issue: when using keyboard to navigate to navigation
buttons (back, forward, and reload) in screencast, they are not focused.

Change-Id: Iac23a5a05065207f706c03f2c19ebd182e296085