Add css rules for navigation buttons when focused in screencast.
Accessibility issue: when using keyboard to navigate to navigation
buttons (back, forward, and reload) in screencast, they are not focused.

We noticed that when the focus is supposed to be on the said buttons,
the button has focus but there is no visual indication. We add a css
rule so that the focus is shown the same way as being hovered over.

Image before: (from navigation url, use shift+tab, to reload or back
buttons. Button does not show focus.)

https://imgur.com/jG66wY6

Images after: (same as above, but the button has an outline color
indicating focus)

https://imgur.com/bMA4zs0




Change-Id: Iac23a5a05065207f706c03f2c19ebd182e296085