Display names for Workers in the Performance panel (if they have them)
Users can give workers a name via the options param when they construct
it. We didn't display this previously on the performance panel, although
we already do in the sources panel.

The user-supplied name is stored on the target, so we need to look up
the target for the worker. The profiling data itself comes from tracing
and doesn't contain the name.

We keep the URL in the output as this is still useful.

User code: https://imgur.com/a/5Y2RmQg
After: https://imgur.com/ZDAccs6

Fixed: 938866

Change-Id: I0e815daef1ab8ae56d0856d73ee673d53724d483