Update patch set 3
Patch Set 3:

Since this will be a separate target, I think it would make sense to put this in `/inspector_overlay/` instead, rather than in `front_end/`. WDYT? That way, we still treat it properly as a separate implementation (similar to how we treat it separately now in the backend) and would also make it more apparent that it can't import code from `front_end/`.

Also, for simplicity, I think it is better to check in the GRD file hardcoded, since it will only have a couple files anyways. You can then use the `copy()` action in Ninja to copy files to a particular location. We are in the process of removing the copy logic from `build_release_applications.py`, so I think it is best to use `copy()` right away in this script.

All in all, that would allow you to remove some complexity from the Python build script and hopefully make it only a call to Rollup. We can then later look into another Ninja action to execute Rollup, so that we can unify the old Rollup logic in `build_release_applications.py` and this script. However, that is suitable for a follow-up CL and not required here.

Patch-set: 3
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>