Use Ninja copy action instead of build_release_applications.py to copy HTML entrypoints
Previously, we were storing the information about the HTML entrypoints
in the `{entrypoint}.json`, so that `build_release_applications.py`
would copy these files over to the `resources/inspector` directory.

Rather than using a Python script for this, we can use the Ninja
copy action. This allows us to enforce that the correct outputs
are passed to `generate_devtools_grd`.

For legacy reasons, we can't apply the same solution to the
integration_test_runner.html. It will therefore be fixed in a
follow-up CL.

A future change can use the target_outputs of
`front_end_html_entrypoints` to automatically generate the GRD
files as well.

R=aerotwist@chromium.org

Bug: 1078770
Change-Id: I8c94c419ace46e881a6920d6d59a3e311d359f2d