Fix protocol dispatcher type generation
The dispatchers must be separate types, as they are implemented
in the SDK models. These models can't implement the *Api interface,
as these are auto-generated in the InspectorBackend.

A follow-up CL will change the Closure type generator to also
understand the new ProtocolProxyApi dispatcher interfaces, to
unblock further typescriptification of the SDK models.

R=jacktfranklin@chromium.org
CC=janscheffler@chromium.org,sigurds@chromium.org

Bug: 1081686, 1011811
Change-Id: Id193def28c426bdce17b9f72c65ccba21efd474e