Update patch set 4
Patch Set 4:

(1 comment)

Adding myself so I can keep updated, as this affects ongoing CSSGrid work.

When this is closer to checkin, should we consider adding a readme/contributing to this folder so devs know how to make/test changes, as is a bit more involved than a normal frontend change?

Patch-set: 4