Update patch set 9
Patch Set 9:

Note if something doesn't work, consider the following: Some of the things that we inherit down from chromium cross-reference each other, e.g. in GN. E.g. possibly: //tools/clang/...
The // refers to the project root, it doesn't matter if it's chromium or v8 or devtools, but the directory structure afterwards needs to be equal for some things to work. It might not be the case with clang, but it certainly is with src/build and some other things from src/tools.

Patch-set: 9
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>