Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> left a comment

Thanks for your comments regarding using a timeout.
I've been bitten by this kind of stuff many times before, you're right it's really important to actually wait for UI changes than for some random time.
I guess I did it here because I do wait for the recording button to change state before proceeding, but I wanted to record a profile for some time. The timeout isn't to wait for the UI to be ready or anything, it's here so the profile does contain a bunch of data.
But you're right nonetheless, and there's another way for me to do this. I could simply wait for the recording message to appear in the UI and maybe check the `time` label in there until it gets over a given value. 
Or maybe I just don't need to wait at all, and stop recording as soon as the message appears. I'll check.
Thanks!

Patch-set: 4