Update patch set 1
Patch Set 1: Code-Review+1 Commit-Queue+1

Mathias and I did a quick investigation as to why ESLint was only warning on `animation-legacy.js` and not on the other `-legacy.js` files. We would have expected ESLint to warn on all of these, since we are explicitly putting these on the global scope.

As it turns out, `Animation` is a pre-existing global in the browser: https://github.com/sindresorhus/globals/blob/f8c183ec1ee238ef274442fe405f0e3d214a14fa/globals.json#L233

ESLint discovered that we overwrite `Animation` and therefore started complaining. Sadly, we can't fix that atm, as both Closure and `Runtime.js` rely on `self.Animation` to exist. Since we are not explicitly using any `window.Animation` API in DevTools itself, we are okay for now. However, if we later want to adopt such an API, we can't do so until we rename `front_end/animation` to a different name.

All that said, the linter rule already caught an issue in the ElementsTestRunner, so this is worthwhile to add for now. It would be great if we can figure out if we can make the rule also warn on other global usages, but punting that to a follow-up.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1
Label: Commit-Queue=+1