Fix nesting of ts_library outputs
For TypeScript-authored files, we would create the wrong output
in the `resources/inspector` directory. While `ts_library.py`
would copy to the correct location, `typescript.gni` disagreed.

The culprit is that we were only taking 1 directory from the
`target_gen_dir`, but this could be arbitrarily many. Instead,
we now compare to the full path of the generated front_end
directory.

To make sure that this works in a standalone and Chromium checkout,
we have to differentiate on `_devtools_location`. I also cleaned
that logic up by always prepending a slash at the end of a
Chromium checkout. As such, you can string interpolate without
having to worry about the extra space.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Bug: 1011811
Change-Id: I186396fe7a5001dc2930a37562640feedb6291a5