Remove tsbuildinfo determinism monkeypatch
This essentially reverts https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2091448

We are now using TypeScript 3.9.2, which includes the fixes to
make `.tsbuildinfo` files deterministic. I confirmed this by
updating our test suite to include the new expected tsbuildinfo
file and confirmed that over a couple of runs it remains stable.

We should monitor the deterministic builder in Chromium to verify
that this is indeed fully fixed now. Moreover, we should submit
this next week after branchpoint, just to be safe.

R=jacktfranklin@chromium.org

Bug: 1054494
Change-Id: I6fcd8b406cbc5cc39f95ab79967b0989ef61e517