Add --chrome-binary argument to unittest script
This CL adds an optional --chrome-binary argument to the unittest.py
script to use a specific Chrome binary to run unit tests against.
When no argument is specified, the script defaults to the downloaded
Chrome binary from `gclient sync` (same behavior as before).

We'd like to have the option of running unit tests on a Microsoft Edge
executable instead of the default downloaded Chrome executable.

Change-Id: Ice72f129086bcc1da56efa118280532cdbb25fd3