Update patch set 3
Patch Set 3: Commit-Queue+1

(6 comments)

Addressed comments.

As mentioned previously, this is to seed the list of owners, not to chisel them in stone.

Leo: I applied some of your suggestions. For others I think it's hard to argue that they already meet the bar. For reference, the lowest bar I applied for some subdirectories is 2 commits/reviews in the past six months. That is already a very liberal interpretation of Chromium's ownership guidelines. We can revisit in the future if this affects velocity or if the contributions have increased, wdyt?

Patch-set: 3
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1